Final Touches Needed

+++ Lab-Sheet-04/03-multiply.c
@@ -1,16 +1,23 @@
 #include<stdio.h>
 
-int main() {
+int main(){
+
 	int a;
 	float b,c;
+	
 	printf("Enter two numbers first int and second float");
 	scanf("%d%f",&a,&b);
+
 	c = mul(a,b);
+
 	printf("%f is the multiplied value");
 }
-float mul(int a,float b){
+
+int mul(int a,float b){
 	float f;
+
 	f = a*b;
+
 	return f;
 }
